# finders-keepers-api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/admin/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/admin/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/admin/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/admin/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/addresses": {
      "post": {
        "operationId": "AddressesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Addresses"
        ]
      },
      "get": {
        "operationId": "AddressesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Addresses"
        ]
      }
    },
    "/addresses/customer/{customerId}": {
      "get": {
        "operationId": "AddressesController_findByCustomer",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Addresses"
        ]
      }
    },
    "/addresses/{id}": {
      "get": {
        "operationId": "AddressesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Addresses"
        ]
      },
      "patch": {
        "operationId": "AddressesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Addresses"
        ]
      },
      "delete": {
        "operationId": "AddressesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Addresses"
        ]
      }
    },
...
    "/categories": {
      "get": {
        "operationId": "CategoriesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ]
      },
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Categories"
        ]
      }
    },
    "/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ]
      },
      "patch": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Categories"
        ]
      }
    },
...
    "/cart": {
      "get": {
        "operationId": "CartController_getCart",
        "parameters": [
          {
            "name": "x-guest-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Cart"
        ]
      }
    },
    "/cart/me": {
      "get": {
        "operationId": "CartController_getCustomerCart",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Cart"
        ]
      }
    },
    "/cart/items": {
      "post": {
        "operationId": "CartController_addItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Cart"
        ]
      }
    },
    "/cart/me/items": {
      "post": {
        "operationId": "CartController_addCustomerItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Cart"
        ]
      }
    },
    "/cart/items/{id}": {
      "patch": {
        "operationId": "CartController_updateItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-guest-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Cart"
        ]
      },
      "delete": {
        "operationId": "CartController_removeItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-guest-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Cart"
        ]
      }
    },
    "/cart/me/items/{id}": {
      "patch": {
        "operationId": "CartController_updateCustomerItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Cart"
        ]
      },
      "delete": {
        "operationId": "CartController_removeCustomerItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Cart"
        ]
      }
    },
    "/cart/clear": {
      "delete": {
        "operationId": "CartController_clearCart",
        "parameters": [
          {
            "name": "x-guest-token",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Cart"
        ]
      }
    },
    "/cart/me/clear": {
      "delete": {
        "operationId": "CartController_clearCustomerCart",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Cart"
        ]
      }
    },
...
  },
  "info": {
    "title": "Finders Keepers API",
    "description": "Ecommerce backend for Finders Keepers LB",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@finderskeeperslb.com"
          },
          "password": {
            "type": "string",
            "example": "Admin@123456"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "CreateAddressDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "example": "customer-id-here"
          },
          "label": {
            "type": "string",
            "example": "Home"
          },
          "fullName": {
            "type": "string",
            "example": "John Doe"
          },
          "phone": {
            "type": "string",
            "example": "+96170123456"
          },
          "country": {
            "type": "string",
            "example": "Lebanon"
          },
          "city": {
            "type": "string",
            "example": "Beirut"
          },
          "area": {
            "type": "string",
            "example": "Hamra"
          },
          "street": {
            "type": "string",
            "example": "Main Street"
          },
          "building": {
            "type": "string",
            "example": "Building A"
          },
          "floor": {
            "type": "string",
            "example": "3"
          },
          "apartment": {
            "type": "string",
            "example": "12"
          },
          "notes": {
            "type": "string",
            "example": "Call before delivery"
          },
          "isDefault": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "customerId"
        ]
      },
      "UpdateAddressDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "example": "customer-id-here"
          },
          "label": {
            "type": "string",
            "example": "Home"
          },
          "fullName": {
            "type": "string",
            "example": "John Doe"
          },
          "phone": {
            "type": "string",
            "example": "+96170123456"
          },
          "country": {
            "type": "string",
            "example": "Lebanon"
          },
          "city": {
            "type": "string",
            "example": "Beirut"
          },
          "area": {
            "type": "string",
            "example": "Hamra"
          },
          "street": {
            "type": "string",
            "example": "Main Street"
          },
          "building": {
            "type": "string",
            "example": "Building A"
          },
          "floor": {
            "type": "string",
            "example": "3"
          },
          "apartment": {
            "type": "string",
            "example": "12"
          },
          "notes": {
            "type": "string",
            "example": "Call before delivery"
          },
          "isDefault": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Clothing"
          },
          "slug": {
            "type": "string",
            "example": "clothing"
          },
          "description": {
            "type": "string"
          },
          "imageId": {
            "type": "string"
          },
          "parentId": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "sortOrder": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Clothing"
          },
          "slug": {
            "type": "string",
            "example": "clothing"
          },
          "description": {
            "type": "string"
          },
          "imageId": {
            "type": "string"
          },
          "parentId": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "sortOrder": {
            "type": "number",
            "example": 0
          }
        }
      },
...
      "AddCartItemDto": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "example": 1
          },
          "guestToken": {
            "type": "string",
            "example": "guest-session-token"
          }
        },
        "required": [
          "variantId",
          "quantity"
        ]
      },
      "UpdateCartItemDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "quantity"
        ]
      },
...
    }
  }
}